chore: move base-time-entity
