+++ src/main/java/org/swmaestro/repl/gifthub/auth/entity/Member.java
@@ -1,7 +1,17 @@
 package org.swmaestro.repl.gifthub.auth.entity;
 
-import jakarta.persistence.*;
-import lombok.*;
+import org.swmaestro.repl.gifthub.util.BaseTimeEntity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Entity
 @Getter
