+++ src/main/java/org/swmaestro/repl/gifthub/auth/entity/RefreshToken.java
@@ -1,11 +1,16 @@
 package org.swmaestro.repl.gifthub.auth.entity;
 
-import jakarta.persistence.*;
+import org.swmaestro.repl.gifthub.util.BaseTimeEntity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
 
 @Entity
 @Getter
